backend/requirements.txt
# FastAPI & server
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Database
sqlalchemy>=2.0
asyncpg>=0.30
alembic>=1.13
greenlet>=3.0

# Redis & Jobs
rq>=1.15
rq-scheduler>=0.13
redis>=5.0

# Google OAuth & APIs
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0

# Security
cryptography>=41.0
itsdangerous>=2.0

# Stripe
stripe>=8.0

# LLM
# Uses `httpx` to call OpenAI Responses API.

# HTTP client
httpx>=0.27

# Date handling
python-dateutil>=2.8

